Refactor sepset finding and simplify path blocking logic
Simplified implementations of sepset finding by removing redundant parameters and unused methods. Introduced `RecursiveBlocking` class to encapsulate and streamline path-blocking logic, enhancing modularity and maintainability.
