+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/SepsetFinder.java
@@ -29,10 +29,9 @@ public class SepsetFinder {
      * @param containing the set of nodes that must be contained in the sepset (optional)
      * @param test       the independence test to use
      * @param depth      the depth of the search
-     * @param order      An order of the nodes in the graph, used for some implementations.
      * @return the sepset containing the greedy test for variables x and y, or null if no sepset is found
      */
-    public static Set<Node> findSepsetSubsetOfAdjxOrAdjy(Graph graph, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth, List<Node> order) {
+    public static Set<Node> findSepsetSubsetOfAdjxOrAdjy(Graph graph, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth) {
         List<Node> adjx = graph.getAdjacentNodes(x);
         List<Node> adjy = graph.getAdjacentNodes(y);
         adjx.remove(y);
@@ -99,81 +98,15 @@ public class SepsetFinder {
      * node combinations from the union of adjacent nodes of x and y in the graph.
      *
      * @param graph      The graph in which the nodes and their adjacency relationships are defined.
-     * @param x          The first node for which the sepset is being determined.
-     * @param y          The second node for which the sepset is being determined.
-     * @param containing A specified subset of nodes that the resulting sepset must contain.
-     * @param test       The independence test to verify conditional independence between nodes.
-     * @param depth      The maximum allowable size for subsets to consider during the search for the sepset.
-     * @param order      A list representing a specific ordering of nodes, which may influence the valid sepset.
-     * @return A set of nodes representing the sepset containing the given subset, or null if no such set is found.
-     */
-    public static Set<Node> getSepsetContainingGreedySubsetUnion(Graph graph, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth, List<Node> order) {
-        List<Node> adjx = graph.getAdjacentNodes(x);
-        List<Node> adjy = graph.getAdjacentNodes(y);
-
-        Set<Node> union = new HashSet<>(adjx);
-        union.addAll(adjy);
-        union.remove(x);
-        union.remove(y);
-
-        List<Node> unionList = new ArrayList<>(union);
-
-        union.removeIf(node -> node.getNodeType() == NodeType.LATENT);
-        union.removeIf(node -> node.getNodeType() == NodeType.LATENT);
-
-        List<List<Integer>> choices = getChoices(unionList, depth);
-
-        // Parallelize processing for adjx
-        Set<Node> sepset = choices.parallelStream()
-                .map(choice -> combination(choice, unionList)) // Generate combinations in parallel
-                .filter(subset -> subset.containsAll(containing)) // Filter combinations that don't contain 'containing'
-                .filter(subset -> {
-                    try {
-//                        if (order != null) {
-//                            Node _y = order.indexOf(x) < order.indexOf(y) ? y : x;
-//
-//                            for (Node node : subset) {
-//                                if (order.indexOf(node) > order.indexOf(_y)) {
-//                                    return false;
-//                                }
-//                            }
-//                        }
-
-                        return separates(x, y, subset, test);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                }) // Further filter by separating sets
-                .findFirst() // Return the first matching subset
-                .orElse(null);
-
-        if (sepset != null) {
-            return sepset;
-        }
-
-        return sepset;
-    }
-
-
-    /**
-     * Identifies a separating set (sepset) containing a given subset of nodes between two nodes x and y in a graph
-     * using a greedy approach and subsets of (adj(x) U adu(y)) \ {x, y}. The method applies constraints such as
-     * independence testing, specified node ordering, and optional filtering of certain node types.
-     * <p>
-     * This method mainly focuses on finding a feasible separating set under the given constraints by iterating through
-     * node combinations from the union of adjacent nodes of x and y in the graph.
-     *
-     * @param graph      The graph in which the nodes and their adjacency relationships are defined.
      * @param cpdag      The CDPDAG.
      * @param x          The first node for which the sepset is being determined.
      * @param y          The second node for which the sepset is being determined.
      * @param containing A specified subset of nodes that the resulting sepset must contain.
      * @param test       The independence test to verify conditional independence between nodes.
      * @param depth      The maximum allowable size for subsets to consider during the search for the sepset.
-     * @param order      A list representing a specific ordering of nodes, which may influence the valid sepset.
      * @return A set of nodes representing the sepset containing the given subset, or null if no such set is found.
      */
-    public static Set<Node> getSepsetContainingGreedySubsetMb(Graph graph, Graph cpdag, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth, List<Node> order) {
+    public static Set<Node> getSepsetContainingGreedySubsetMb(Graph graph, Graph cpdag, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth) {
         List<Node> mbx = new ArrayList<>(cpdag.paths().markovBlanket(x));
 
         if (("A".equals(x.getName()) && "E".equals(y.getName()))
@@ -330,16 +263,15 @@ public class SepsetFinder {
     /**
      * Returns the sepset containing the minimum p-value for the given variables x and y.
      *
-     * @param graph      the graph representing the network
-     * @param x          the first node
-     * @param y          the second node
-     * @param containing the set of nodes to be excluded from the sepset
-     * @param test       the independence test to use for calculating the p-value
-     * @param depth      the depth of the search for the sepset
+     * @param graph the graph representing the network
+     * @param x     the first node
+     * @param y     the second node
+     * @param test  the independence test to use for calculating the p-value
+     * @param depth the depth of the search for the sepset
      * @return the sepset containing the minimum p-value, or null if no sepset is found
      * @throws InterruptedException if any
      */
-    public static Set<Node> getSepsetContainingMinPHybrid(Graph graph, Node x, Node y, Set<Node> containing, IndependenceTest test, int depth) throws InterruptedException {
+    public static Set<Node> getSepsetContainingMinPHybrid(Graph graph, Node x, Node y, IndependenceTest test, int depth) throws InterruptedException {
         List<Node> adjx = graph.getAdjacentNodes(x);
         List<Node> adjy = graph.getAdjacentNodes(y);
         adjx.remove(y);
@@ -418,96 +350,7 @@ public class SepsetFinder {
      */
     public static Set<Node> blockPathsRecursively(Graph graph, Node x, Node y, Set<Node> containing, Set<Node> notFollowed,
                                                   int maxPathLength) throws InterruptedException {
-        return blockPathsRecursivelyVisit(graph, x, y, containing, notFollowed, graph.paths().getDescendantsMap(), maxPathLength);
-    }
-
-    private static Set<Node> blockPathsRecursivelyVisit(Graph graph, Node x, Node y, Set<Node> containing,
-                                                        Set<Node> notFollowed, Map<Node, Set<Node>> ancestorMap, int maxPathLength)
-            throws InterruptedException {
-        if (x == y) {
-            return null;
-        }
-
-        Set<Node> z = new HashSet<>(containing);
-
-        Set<Node> path = new HashSet<>();
-        path.add(x);
-
-        for (Node b : graph.getAdjacentNodes(x)) {
-            if (Thread.currentThread().isInterrupted()) {
-                return null;
-            }
-
-            findPathToTarget(graph, x, b, y, path, z, maxPathLength, notFollowed, ancestorMap);
-        }
-
-        return z;
-    }
-
-    private static List<Node> getReachableNodes(Graph graph, Node a, Node b, Set<Node> z, Map<Node, Set<Node>> ancestorMap) {
-        List<Node> passNodes = new ArrayList<>();
-
-        for (Node c : graph.getAdjacentNodes(b)) {
-            if (c == a) {
-                continue;
-            }
-
-            if (reachable(graph, a, b, c, z, ancestorMap)) {
-                passNodes.add(c);
-            }
-        }
-
-        return passNodes;
-    }
-
-    private static List<Node> getReachableNodes(Graph graph, Node a, Node b, Set<Node> z) {
-        List<Node> passNodes = new ArrayList<>();
-
-        for (Node c : graph.getAdjacentNodes(b)) {
-            if (c == a) {
-                continue;
-            }
-
-            if (reachable(graph, a, b, c, z)) {
-                passNodes.add(c);
-            }
-        }
-
-        return passNodes;
-    }
-
-    private static boolean reachable(Graph graph, Node a, Node b, Node c, Set<Node> z,
-                                     Map<Node, Set<Node>> ancestors) {
-        boolean collider = graph.isDefCollider(a, b, c);
-
-        if ((!collider || graph.isUnderlineTriple(a, b, c)) && !z.contains(b)) {
-            return true;
-        }
-
-        if (ancestors == null) {
-            return collider && graph.paths().isAncestorOfAnyZ(b, z);
-        } else {
-            boolean ancestor = false;
-
-            for (Node _z : ancestors.get(b)) {
-                if (z.contains(_z)) {
-                    ancestor = true;
-                    break;
-                }
-            }
-
-            return collider && ancestor;
-        }
-    }
-
-    private static boolean reachable(Graph graph, Node a, Node b, Node c, Set<Node> z) {
-        boolean collider = graph.isDefCollider(a, b, c);
-
-        if ((!collider || graph.isUnderlineTriple(a, b, c)) && !z.contains(b)) {
-            return true;
-        }
-
-        return collider && graph.paths().isAncestorOfAnyZ(b, z);
+        return RecursiveBlocking.blockPathsRecursively(graph, x, y, containing, notFollowed, maxPathLength);
     }
 
     /**
@@ -558,129 +401,11 @@ public class SepsetFinder {
 
             // Iterate over adjacent nodes to find potential paths.
             for (Node b : graph.getAdjacentNodes(x)) {
-                findPathToTarget(graph, x, b, y, path, z, maxPathLength, notFollowed, ancestorMap);
+                RecursiveBlocking.findPathToTarget(graph, x, b, y, path, z, maxPathLength, notFollowed, ancestorMap);
             }
         } while (!previousZ.equals(z)); // Repeat if the set z changes.
 
         return z;
-    }
-
-    /**
-     * Finds a path from node a to node b that can be blocked by conditioning on a set of nodes z. The method returns
-     * true if the path can be blocked, and false otherwise.
-     * <p>
-     * The side effects of this method are changes to z and colliders; this method is private, and the public methods
-     * that call it are responsible for handling these side effects.
-     *
-     * @param graph         The graph containing the nodes.
-     * @param a             The first node in the pair.
-     * @param b             The second node in the pair.
-     * @param y             The target node.
-     * @param path          The current path.
-     * @param z             The set of nodes that can block the path. This is a set of conditioning nodes that is being
-     *                      built.
-     * @param maxPathLength The maximum length of the paths to consider.
-     * @param notFollowed   A set of nodes that should not be followed along paths.
-     * @return True if the path can be blocked, false otherwise.
-     */
-    private static Blockable findPathToTarget(Graph graph, Node a, Node b, Node y, Set<Node> path, Set<Node> z,
-                                              int maxPathLength, Set<Node> notFollowed, Map<Node, Set<Node>> ancestorMap)
-            throws InterruptedException {
-        if (Thread.currentThread().isInterrupted()) {
-            return Blockable.INDETERMINATE;
-        }
-
-        if (b == y) {
-            return Blockable.UNBLOCKABLE;
-        }
-
-        if (path.contains(b)) {
-            return Blockable.UNBLOCKABLE;
-        }
-
-        path.add(b);
-
-        if (maxPathLength != -1) {
-            if (path.size() > maxPathLength) {
-                return Blockable.INDETERMINATE;
-            }
-        }
-
-        // If b is latent, we cannot condition on it. If z already contains b, we know we've already conditioned on
-        // it, so there's no point considering further whether to condition on it or now.
-        if (b.getNodeType() == NodeType.LATENT || z.contains(b)) {
-            List<Node> passNodes = getReachableNodes(graph, a, b, z, ancestorMap);
-            passNodes.removeAll(notFollowed);
-
-            for (Node c : passNodes) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                Blockable blockable = findPathToTarget(graph, b, c, y, path, z, maxPathLength, notFollowed, ancestorMap);
-
-                if (blockable == Blockable.UNBLOCKABLE || blockable == Blockable.INDETERMINATE) {
-                    return Blockable.UNBLOCKABLE;
-                }
-            }
-
-            path.remove(b);
-            return Blockable.BLOCKED; // blocked.
-        } else {
-
-            // We're going to look to see whether the path to y has already been blocked by z. If it has, we can
-            // stop here. If it hasn't, we'll see if we can block it by conditioning also on b. If it can't be
-            // blocked either way, well, then, it just can't be blocked.
-            boolean blockable1 = true;
-
-            List<Node> passNodes = getReachableNodes(graph, a, b, z, ancestorMap);
-            passNodes.removeAll(notFollowed);
-
-            for (Node c : passNodes) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                Blockable blockType = findPathToTarget(graph, b, c, y, path, z, maxPathLength, notFollowed, ancestorMap);
-
-                if (blockType == Blockable.UNBLOCKABLE || blockType == Blockable.INDETERMINATE) {
-                    blockable1 = false;
-                    break;
-                }
-            }
-
-            if (blockable1) {
-                path.remove(b);
-                return Blockable.BLOCKED;
-            }
-
-            z.add(b);
-
-            boolean blockable2 = true;
-            passNodes = getReachableNodes(graph, a, b, z, ancestorMap);
-            passNodes.removeAll(notFollowed);
-
-            for (Node c : passNodes) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                Blockable blackable = findPathToTarget(graph, b, c, y, path, z, maxPathLength, notFollowed, ancestorMap);
-
-                if (blackable == Blockable.UNBLOCKABLE || blackable == Blockable.INDETERMINATE) {
-                    blockable2 = false;
-                    break;
-                }
-            }
-
-            if (blockable2) {
-                path.remove(b);
-                return Blockable.BLOCKED;
-            }
-
-            path.remove(b);
-            return Blockable.UNBLOCKABLE;
-        }
     }
 
     private static @NotNull List<List<Integer>> getChoices(List<Node> adjx, int depth) {
@@ -715,7 +440,7 @@ public class SepsetFinder {
     }
 
     private static double getPValue(Node x, Node y, Set<Node> combination, IndependenceTest test) {
-        double pValue = 0;
+        double pValue;
         try {
             pValue = test.checkIndependence(x, y, combination).getPValue();
         } catch (Exception e) {
@@ -758,7 +483,7 @@ public class SepsetFinder {
                 continue;
             }
 
-            Node node = path.get(path.size() - 1);
+            Node node = path.getLast();
 
             if (node == y) continue;
 
@@ -906,274 +631,4 @@ public class SepsetFinder {
 
         path.remove(a);
     }
-
-    /**
-     * Calculates the sepset path blocking out-of operation for a given pair of nodes in a graph. This method searches
-     * for m-connecting paths out of x and y, and then tries to block these paths by conditioning on definite
-     * noncollider nodes. If all paths are blocked, the method returns the sepset; otherwise, it returns null. The
-     * length of the paths to consider can be limited by the maxLength parameter, and the depth of the final sepset can
-     * be limited by the depth parameter. When increasing the considered path length does not yield any new paths, the
-     * search is terminated early.
-     *
-     * @param mpdag              The graph representing the Markov equivalence class that contains the nodes.
-     * @param x                  The first node in the pair.
-     * @param y                  The second node in the pair.
-     * @param test               The independence test object to use for checking independence.
-     * @param maxLength          The maximum length of the paths to consider. If set to a negative value or a value
-     *                           greater than the number of nodes minus one, it is adjusted accordingly.
-     * @param depth              The maximum depth of the final sepset. If set to a negative value, no limit is
-     *                           applied.
-     * @param allowSelectionBias A boolean flag indicating whether to allow selection bias.
-     * @param blacklist          The set of nodes to blacklist.
-     * @return The sepset if independence holds, otherwise null.
-     * @throws InterruptedException if any
-     */
-    public static Set<Node> getSepsetPathBlockingOutOfX(Graph mpdag, Node x, Node y, IndependenceTest test,
-                                                        int maxLength, int depth, boolean allowSelectionBias,
-                                                        Set<Node> blacklist) throws InterruptedException {
-        int maxLength1 = maxLength;
-        if (maxLength1 < 0 || maxLength1 > mpdag.getNumNodes() - 1) {
-            maxLength1 = mpdag.getNumNodes() - 1;
-        }
-
-        Set<Node> conditioningSet = new HashSet<>();
-        Set<Node> couldBeColliders = new HashSet<>();
-
-        Set<List<Node>> paths = bfsAllPathsOutOfX(mpdag, conditioningSet, couldBeColliders, blacklist, maxLength1, x, y, allowSelectionBias);
-
-        List<Node> couldBeCollidersList = new ArrayList<>(couldBeColliders);
-        conditioningSet.removeAll(couldBeColliders);
-
-        SublistGenerator generator = new SublistGenerator(couldBeCollidersList.size(), depth);
-        int[] choice;
-
-        while ((choice = generator.next()) != null) {
-            Set<Node> sepset = new HashSet<>();
-
-            for (int k : choice) {
-                sepset.add(couldBeCollidersList.get(k));
-            }
-
-            sepset.addAll(conditioningSet);
-
-            if (depth != -1 && sepset.size() > depth) {
-                continue;
-            }
-
-            sepset.remove(y);
-
-            if (test.checkIndependence(x, y, sepset).isIndependent()) {
-                Set<Node> _z = new HashSet<>(sepset);
-                boolean removed;
-
-                do {
-                    removed = false;
-
-                    for (Node w : new HashSet<>(_z)) {
-                        Set<Node> __z = new HashSet<>(_z);
-
-                        __z.remove(w);
-
-                        if (test.checkIndependence(x, y, __z).isIndependent()) {
-                            removed = true;
-                            _z = __z;
-                        }
-                    }
-                } while (removed);
-
-                sepset = new HashSet<>(_z);
-
-//                if (verbose) {
-//                    TetradLogger.getInstance().log("\n\tINDEPENDENCE HOLDS!: " + LogUtilsSearch.independenceFact(x, y, sepset));
-//                }
-
-                return sepset;
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Performs a breadth-first search to find all paths out of a specific node in a graph, considering certain
-     * conditions and constraints.
-     *
-     * @param graph              the graph to search
-     * @param conditionSet       the set of nodes that need to be conditioned on
-     * @param couldBeColliders   the set of nodes that could potentially be colliders
-     * @param blacklist          the set of nodes to exclude from the search
-     * @param maxLength          the maximum length of the paths (-1 for unlimited)
-     * @param x                  the starting node
-     * @param y                  the destination node
-     * @param allowSelectionBias flag to indicate whether to allow selection bias in path selection
-     * @return a set of all paths that satisfy the conditions and constraints
-     * @throws IllegalArgumentException if the conditioning set is null
-     */
-    public static Set<List<Node>> bfsAllPathsOutOfX(Graph graph, Set<Node> conditionSet, Set<Node> couldBeColliders,
-                                                    Set<Node> blacklist, int maxLength, Node x, Node y, boolean allowSelectionBias)
-            throws InterruptedException {
-        Set<List<Node>> allPaths = new HashSet<>();
-        Queue<List<Node>> queue = new LinkedList<>();
-        queue.add(Collections.singletonList(x));
-
-        if (conditionSet == null) {
-            throw new IllegalArgumentException("Conditioning set cannot be null.");
-        }
-
-        while (!queue.isEmpty()) {
-            if (Thread.currentThread().isInterrupted()) {
-                throw new InterruptedException();
-            }
-
-            List<Node> path = queue.poll();
-
-            if (maxLength != -1 && path.size() > maxLength) {
-                continue;
-            }
-
-            Node node = path.get(path.size() - 1);
-
-            if (path.size() < 2) {
-                allPaths.add(path);
-            }
-
-            if (path.size() >= 2 && graph.paths().isMConnectingPath(path, conditionSet, allowSelectionBias)) {
-                allPaths.add(path);
-            }
-
-            for (Node z3 : graph.getAdjacentNodes(node)) {
-                if (Thread.currentThread().isInterrupted()) {
-                    throw new InterruptedException();
-                }
-
-                if (!path.contains(z3)) {
-                    List<Node> newPath = new ArrayList<>(path);
-                    newPath.add(z3);
-
-                    if (newPath.size() - 1 == 1) {
-                        queue.add(newPath);
-                    }
-
-                    // If the path is of at least length 1, and the last two nodes on the path form a noncollider
-                    // with 'adjacent', we need to block these noncolliders first by conditioning on node.
-                    if (newPath.size() - 1 > 1) {
-                        Node z1 = newPath.get(newPath.size() - 3);
-                        Node z2 = newPath.get(newPath.size() - 2);
-
-                        if (!graph.isDefCollider(z1, z2, z3)) {
-//                            if (blacklist.contains(z2)) {
-//                                continue;
-//                            }
-
-                            blockPath(newPath, graph, conditionSet, couldBeColliders, blacklist, x, y, true);
-
-                            if (graph.paths().isMConnectingPath(newPath, conditionSet, allowSelectionBias)) {
-                                queue.add(newPath);
-                            }
-                        }
-                    }
-                }
-            }
-
-            for (Node z3 : graph.getAdjacentNodes(node)) {
-                if (!path.contains(z3)) {
-                    List<Node> newPath = new ArrayList<>(path);
-                    newPath.add(z3);
-
-                    if (newPath.size() - 1 == 1) {
-                        queue.add(newPath);
-                    }
-
-                    // If the path is of at least length 1, and the last two nodes on the path form a noncollider
-                    // with 'adjacent', we need to block these noncolliders first by conditioning on node.
-                    if (newPath.size() - 1 > 1) {
-                        Node z1 = newPath.get(newPath.size() - 3);
-                        Node z2 = newPath.get(newPath.size() - 2);
-
-                        if (graph.isDefCollider(z1, z2, z3)) {
-                            blockPath(newPath, graph, conditionSet, couldBeColliders, blacklist, x, y, true);
-
-                            if (graph.paths().isMConnectingPath(newPath, conditionSet, allowSelectionBias)) {
-                                queue.add(newPath);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        return allPaths;
-    }
-
-    /**
-     * Tries to block the given path is blocked by conditioning on definite noncollider nodes. Return true if the path
-     * is blocked, false otherwise.
-     *
-     * @param path             the path to check
-     * @param graph            the MPDAG graph to analyze
-     * @param conditioningSet  the set of nodes to condition on; this may be modified
-     * @param couldBeColliders the set of nodes that could be colliders; this may be modified
-     * @param y                the second node
-     * @param verbose          whether to print trace information
-     */
-    private static void blockPath(List<Node> path, Graph graph, Set<Node> conditioningSet, Set<Node> couldBeColliders, Set<Node> blacklist,
-                                  Node x, Node y, boolean verbose) {
-
-        for (int n = 1; n < path.size() - 1; n++) {
-            Node z1 = path.get(n - 1);
-            Node z2 = path.get(n);
-            Node z3 = path.get(n + 1);
-
-            if (z2.getNodeType() == NodeType.LATENT) {
-                continue;
-            }
-
-            if (z1.getNodeType().equals(NodeType.LATENT) || z3.getNodeType().equals(NodeType.LATENT)) {
-                continue;
-            }
-
-            if (z1 == x && z3 == y && graph.isDefCollider(z1, z2, z3)) {
-                blacklist.add(z2);
-                break;
-            }
-
-            if (!graph.isDefCollider(z1, z2, z3)) {
-                if (conditioningSet.contains(z2)) {
-//                    if (verbose) {
-//                        TetradLogger.getInstance().log("This " + path + "--is already blocked by " + z2);
-//                    }
-
-                    conditioningSet.removeAll(blacklist);
-                    addCouldBeCollider(z1, z2, z3, path, graph, couldBeColliders, verbose);
-                }
-
-                conditioningSet.add(z2);
-                conditioningSet.removeAll(blacklist);
-
-//                if (verbose) {
-//                    TetradLogger.getInstance().log("Blocking " + path + " with noncollider " + z2);
-//                }
-
-                // If this noncollider is adjacent to the endpoints (i.e. is covered), we note that
-                // it could be a collider. We will need to either consider this to be a collider or
-                // a noncollider below.
-                addCouldBeCollider(z1, z2, z3, path, graph, couldBeColliders, verbose);
-                break;
-            }
-        }
-    }
-
-    private static void addCouldBeCollider(Node z1, Node z2, Node z3, List<Node> path, Graph mpdag,
-                                           Set<Node> couldBeColliders, boolean verbose) {
-        if (mpdag.isAdjacentTo(z1, z3)) {
-            couldBeColliders.add(z2);
-
-//            if (verbose) {
-//                TetradLogger.getInstance().log("Noting that " + z2 + " could be a collider on " + path);
-//            }
-        }
-    }
-
-    public static enum Blockable {BLOCKED, UNBLOCKABLE, INDETERMINATE}
-
 }
