+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/utils/SepsetsGreedy.java
@@ -74,7 +74,7 @@ public class SepsetsGreedy implements SepsetProducer {
      * @return The sepset between the two nodes
      */
     public Set<Node> getSepset(Node i, Node k, int depth, List<Node> order) {
-        return SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, new HashSet<>(), this.independenceTest, depth, order);
+        return SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, new HashSet<>(), this.independenceTest, depth);
     }
 
     /**
@@ -89,14 +89,14 @@ public class SepsetsGreedy implements SepsetProducer {
      */
     @Override
     public Set<Node> getSepsetContaining(Node i, Node k, Set<Node> s, int depth) {
-        return SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, s, this.independenceTest, depth, null);
+        return SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, s, this.independenceTest, depth);
     }
 
     /**
      * {@inheritDoc}
      */
     public boolean isUnshieldedCollider(Node i, Node j, Node k, int depth) {
-        Set<Node> set = SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, null, this.independenceTest, depth, null);
+        Set<Node> set = SepsetFinder.findSepsetSubsetOfAdjxOrAdjy(graph, i, k, null, this.independenceTest, depth);
         return set != null && !set.contains(j);
     }
